+++ src/Components/RecentProductList/index.js
@@ -1,0 +1,7 @@
+import React, { Component } from 'react';
+
+export default class index extends Component {
+  render() {
+    return <div></div>;
+  }
+}
